fix: handle list of Upath
An error would be thrown when the path variable is a Sequence[UPath] because there is no .as_uri() method for this case. Instead, all paths in the sequence should have .as_uri() called on them independently, and then the list of uris can be fed to the mechanism for reading the parquet file.